mcp-servers/linkedin-carousel-generator/pyproject.toml
[project]
name = "linkedin-carousel-generator"
version = "0.1.0"
description = "Automated LinkedIn carousel generator from newsletter markdown"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "openai>=1.12.0",
    "img2pdf>=0.5.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "markdown>=3.5",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["linkedin_carousel"]
package-dir = {"" = "src"}
